Refactor to better differentiate lifecycles and phases.
Previously, phases were sort of called lifecycles, but once we
introduced more than one lifecycle things broke down a bit.
