+++ lib/datatypes/environment-delete-result.js
@@ -1,0 +1,9 @@
+class EnvironmentDeleteResult {
+    constructor(status, message, error) {
+        this.status = status;
+        this.message = message;
+        this.error = error;
+    }
+}
+
+module.exports = exports = EnvironmentDeleteResult;
