+++ lib/datatypes/environment-deploy-result.js
@@ -1,0 +1,9 @@
+class EnvironmentDeployResult {
+    constructor(status, message, error) {
+        this.status = status;
+        this.message = message;
+        this.error = error;
+    }
+}
+
+module.exports = exports = EnvironmentDeployResult;
