+++ lib/phases/bind.js
