+++ lib/phases/check.js
