+++ lib/phases/consume-events.js
