+++ lib/phases/produce-events.js
