+++ lib/util/util.js
@@ -3,6 +3,7 @@ const path = require('path');
 const yaml = require('js-yaml');
 const archiver = require('archiver');
 const winston = require('winston');
+const AWS = require('aws-sdk');
 
 exports.readFileSync = function(filePath) {
     try {
@@ -152,4 +153,22 @@ exports.getHandelFileParser = function(handelFile) {
         winston.error(`Invalid deploy spec version: ${handelFile.version}`);
         return null;
     }
+}
+
+/**
+ * Gets the App Context from the deploy spec file
+ */
+exports.createEnvironmentContext = function(handelFile, handelFileParser, environmentName, deployVersion) {
+    try {
+        return handelFileParser.createEnvironmentContext(handelFile, environmentName, deployVersion);
+    }
+    catch(err) {
+        winston.error(`Error while parsing deploy spec: ${err.message}`);
+        return null;
+    }
+}
+
+
+exports.configureAwsSdk = function(accountConfig) {
+    AWS.config.update({region: accountConfig.region});
 }
