+++ test/phases/check-test.js
@@ -1,5 +1,5 @@
 const accountConfig = require('../../lib/util/account-config')(`${__dirname}/../test-account-config.yml`).getAccountConfig();
-const checkPhase = require('../../lib/lifecycle/check');
+const checkPhase = require('../../lib/phases/check');
 const EnvironmentContext = require('../../lib/datatypes/environment-context');
 const ServiceContext = require('../../lib/datatypes/service-context');
 const expect = require('chai').expect;
