+++ test/phases/consume-events-test.js
@@ -1,15 +1,14 @@
 const accountConfig = require('../../lib/util/account-config')(`${__dirname}/../test-account-config.yml`).getAccountConfig();
-const ProduceEventsContext = require('../../lib/datatypes/produce-events-context');
+const ConsumeEventsContext = require('../../lib/datatypes/consume-events-context');
 const DeployContext = require('../../lib/datatypes/deploy-context');
 const ServiceContext = require('../../lib/datatypes/service-context');
 const EnvironmentContext = require('../../lib/datatypes/environment-context');
-const ConsumeEventsContext = require('../../lib/datatypes/consume-events-context');
-const produceEvents = require('../../lib/lifecycle/produce-events');
+const consumeEvents = require('../../lib/phases/consume-events');
 const expect = require('chai').expect;
 const sinon = require('sinon');
 const util = require('../../lib/util/util');
 
-describe('produceEvents module', function() {
+describe('consumeEvents module', function() {
     let sandbox;
 
     beforeEach(function() {
@@ -20,18 +19,17 @@ describe('produceEvents module', function() {
         sandbox.restore();
     });
 
-    describe('produceEvents', function() {
-        it('should execute produceEvents on all services that specify themselves as producers for other services', function() {
+    describe('consumeEvents', function() {
+        it('should execute consumeEvents on all services that are specified as consumers by other services', function() {
             let serviceDeployers = {
                 lambda: {
-                    produceEvents: function(ownServiceContext, ownDeployContext, consumerServiceContext, consumerDeployContext) {
-                        return Promise.reject(new Error("Lambda doesn't produce events"));
-                        
+                    consumeEvents: function(ownServiceContext, ownDeployContext, producerServiceContext, producerDeployContext) {
+                        return Promise.resolve(new ConsumeEventsContext(ownServiceContext, producerServiceContext));
                     }
                 },
                 s3: {
-                    produceEvents: function(ownServiceContext, ownDeployContext, consumerServiceContext, consumerDeployContext) {
-                        return Promise.resolve(new ProduceEventsContext(ownServiceContext,  consumerServiceContext));
+                    consumeEvents: function(ownServiceContext, ownDeployContext, producerServiceContext, producerDeployContext) {
+                        return Promise.reject(new Error("S3 doesn't consume events"));
                     }
                 }
             };
@@ -68,9 +66,9 @@ describe('produceEvents module', function() {
             deployContexts[serviceNameA] = new DeployContext(serviceContextA);
             deployContexts[serviceNameB] = new DeployContext(serviceContextB);
 
-            return produceEvents.produceEvents(serviceDeployers, environmentContext, deployContexts)
-                .then(produceEventsContext => {
-                    expect(produceEventsContext['A->B']).to.be.instanceof(ProduceEventsContext);
+            return consumeEvents.consumeEvents(serviceDeployers, environmentContext, deployContexts)
+                .then(consumeEventsContexts => {
+                    expect(consumeEventsContexts['B->A']).to.be.instanceof(ConsumeEventsContext);
                 });
         });
     });
