+++ test/phases/produce-events-test.js
@@ -1,14 +1,15 @@
 const accountConfig = require('../../lib/util/account-config')(`${__dirname}/../test-account-config.yml`).getAccountConfig();
-const ConsumeEventsContext = require('../../lib/datatypes/consume-events-context');
+const ProduceEventsContext = require('../../lib/datatypes/produce-events-context');
 const DeployContext = require('../../lib/datatypes/deploy-context');
 const ServiceContext = require('../../lib/datatypes/service-context');
 const EnvironmentContext = require('../../lib/datatypes/environment-context');
-const consumeEvents = require('../../lib/lifecycle/consume-events');
+const ConsumeEventsContext = require('../../lib/datatypes/consume-events-context');
+const produceEvents = require('../../lib/phases/produce-events');
 const expect = require('chai').expect;
 const sinon = require('sinon');
 const util = require('../../lib/util/util');
 
-describe('consumeEvents module', function() {
+describe('produceEvents module', function() {
     let sandbox;
 
     beforeEach(function() {
@@ -19,17 +20,18 @@ describe('consumeEvents module', function() {
         sandbox.restore();
     });
 
-    describe('consumeEvents', function() {
-        it('should execute consumeEvents on all services that are specified as consumers by other services', function() {
+    describe('produceEvents', function() {
+        it('should execute produceEvents on all services that specify themselves as producers for other services', function() {
             let serviceDeployers = {
                 lambda: {
-                    consumeEvents: function(ownServiceContext, ownDeployContext, producerServiceContext, producerDeployContext) {
-                        return Promise.resolve(new ConsumeEventsContext(ownServiceContext, producerServiceContext));
+                    produceEvents: function(ownServiceContext, ownDeployContext, consumerServiceContext, consumerDeployContext) {
+                        return Promise.reject(new Error("Lambda doesn't produce events"));
+                        
                     }
                 },
                 s3: {
-                    consumeEvents: function(ownServiceContext, ownDeployContext, producerServiceContext, producerDeployContext) {
-                        return Promise.reject(new Error("S3 doesn't consume events"));
+                    produceEvents: function(ownServiceContext, ownDeployContext, consumerServiceContext, consumerDeployContext) {
+                        return Promise.resolve(new ProduceEventsContext(ownServiceContext,  consumerServiceContext));
                     }
                 }
             };
@@ -66,9 +68,9 @@ describe('consumeEvents module', function() {
             deployContexts[serviceNameA] = new DeployContext(serviceContextA);
             deployContexts[serviceNameB] = new DeployContext(serviceContextB);
 
-            return consumeEvents.consumeEvents(serviceDeployers, environmentContext, deployContexts)
-                .then(consumeEventsContexts => {
-                    expect(consumeEventsContexts['B->A']).to.be.instanceof(ConsumeEventsContext);
+            return produceEvents.produceEvents(serviceDeployers, environmentContext, deployContexts)
+                .then(produceEventsContext => {
+                    expect(produceEventsContext['A->B']).to.be.instanceof(ProduceEventsContext);
                 });
         });
     });
